pyproject.toml
[project]
name = "apna-dukan"
version = "0.1.0"
description = ""
authors = [
    {name = "Akash Kumar",email = "akashkr15082000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery (>=5.6.3,<6.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn[standard] (>=0.46.0,<0.47.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "httpx (>=0.28.1,<0.29.0)",
    "anthropic (>=0.97.0,<0.98.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "weasyprint (>=68.1,<69.0)",
    "boto3 (>=1.42.95,<2.0.0)",
    "razorpay (>=2.0.1,<3.0.0)",
    "openai (>=2.32.0,<3.0.0)",
    "google-genai (>=1.73.1,<2.0.0)",
    "groq (>=1.2.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "apna_dukan", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
